+++ app/domain/secrets/secret_types/secret_base_type.rb
@@ -1,47 +1,56 @@
 module Secrets
   module SecretTypes
     class SecretBaseType
-      def create_secret(policy, resource_id, params, response)
+      include PermissionsHandler
+      include AnnotationsHandler
+
+      def create_secret(policy, secret_id, params, response)
         # Create variable resource
-        variable_resource = ::Resource.create(resource_id: resource_id, owner_id: policy[:resource_id], policy_id: policy[:resource_id])
+        variable_resource = ::Resource.create(resource_id: secret_id, owner_id: policy[:resource_id], policy_id: policy[:resource_id])
 
         # Add annotations
         if params[:annotations].nil?
           response["annotations"] = "[]"
         end
         annotations = add_annotations(params)
-        AnnotationsHandler.create_annotations(variable_resource, policy, annotations)
+        create_annotations(variable_resource, policy, annotations)
+
+        # Add permissions
+        if params[:permissions].nil?
+          response["permissions"] = "[]"
+        else
+          allowed_privilege = %w[read execute update]
+          resources_privileges = validate_permissions(params[:permissions], allowed_privilege)
+          add_permissions(resources_privileges, secret_id, policy)
+        end
 
         # Set secret value
         set_value(variable_resource, params[:value])
-        # remove value field if exists
+
+        # remove value field if exists from the response
         response.delete("value")
 
         # Convert the modified data back to a JSON string
         JSON.generate(response)
       rescue Sequel::UniqueConstraintViolation => e
-        raise Exceptions::RecordExists.new("secret", resource_id)
+        raise Exceptions::RecordExists.new("secret", secret_id)
       end
 
       def set_value(variable_resource, value)
         #No implementation
       end
 
-      def add_annotations(params)
-        # create annotations from secret fields
-        annotations = convert_fields_to_annotations(params)
-        # add secret annotations
-        unless (params[:annotations]).nil?
-          params[:annotations].each do |obj|
-            annotations[obj["name"]] = obj["value"]
-          end
-        end
-
-        annotations
-      end
-
       def convert_fields_to_annotations(params)
         {"conjur/kind" => params[:type]}
+      end
+
+      private
+      def add_annotations(params)
+        annotations = convert_annotations_object(params)
+        # create annotations from secret fields
+        annotations.merge! convert_fields_to_annotations(params)
+
+        annotations
       end
     end
   end